made different types of towers

+++ ChromazoneGame/Assets/Scripts/TowerScript.cs
@@ -74,7 +74,7 @@ public class TowerScript : MonoBehaviour
         //make towerhealthindi be proportional to health
         GameObject canvas = gameObject.transform.GetChild(1).gameObject;
         GameObject HPBar = canvas.transform.GetChild(0).gameObject;
-        HPBar.GetComponent<Image>().fillAmount = Health / maxHealth;
+        HPBar.GetComponent<Image>().fillAmount = (float)Health / (float)maxHealth;
 
         //set the bullet speed and damage
         if (counter <= bulletCooldown || closestEnemy == null)
